SONAR-75 rename i18n bundles to l10n bundles
Article from the W3C :
Localization refers to the adaptation of a product, application or document content to meet the language, cultural and other requirements of a specific target market (a "locale").

Internationalization is the design and development of a product, application or document content that enables easy localization for target audiences that vary in culture, region, or language.